FlatArray: Add 1-16 param constructors
