+++ src/collections-flat-array/Collections.FlatArray/FlatArray.T/FlatArray.Factory.Constructor.cs
@@ -1,5 +1,4 @@
-﻿using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
+﻿using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
 
 namespace System.Collections.Generic;
@@ -17,20 +16,5 @@ partial struct FlatArray<T>
 
         length = source.Length;
         items = InnerArrayHelper.Clone(source);
-    }
-
-    // Initializes an instance in 'as is' mode without any processing and creation of a defensive copy
-    //
-    // Since the invariant of FlatArray implies the empty FlatArray contains null underlying array,
-    // the caller MUST ensure the length is GREATER than zero
-    //
-    // Note: The unused arg is intended to separate this from the public one
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private FlatArray(T[] items, int _)
-    {
-        Debug.Assert(items.Length != default);
-
-        length = items.Length;
-        this.items = items;
     }
 }
